Move functions only used by checkArgs to checkArgs module and add more tests
