+++ test/test_scenario_1.py
@@ -10,7 +10,8 @@ from test import DATA_DIR, TEST_DIR, pushd
 
 from RAiDER.constants import Zenith
 from RAiDER.delay import tropo_delay
-from RAiDER.utilFcns import gdal_open, makeDelayFileNames, modelName2Module
+from RAiDER.utilFcns import gdal_open
+from RAiDER.checkArgs import makeDelayFileNames, modelName2Module
 
 SCENARIO_DIR = os.path.join(TEST_DIR, "scenario_1")
 _RTOL = 1e-4
